+++ RELEASE_NOTES.md
@@ -2,7 +2,7 @@
 
 Changes:
 
-
+  * Fixed an issue where Run Scanner could add more containers to a run than should be allowed
 
 # 0.2.202